# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express MySQL Docker API",
    "version": "1.0.0",
    "description": "ノート・クイズ管理アプリケーションの REST API"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API サーバー"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT トークンを Authorization ヘッダーに設定"
      }
    },
    "schemas": {
...
      "Note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/notes": {
      "get": {
        "tags": [
          "Notes"
        ],
        "summary": "ノート一覧取得",
        "description": "全ノートを updatedAt の降順で取得します。認証不要です。test",
        "responses": {
          "200": {
            "description": "ノート一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Note"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Notes"
        ],
        "summary": "ノート作成",
        "description": "新規ノートを作成します。Bearer認証が必要です。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "タイトル"
                  },
                  "content": {
                    "type": "string",
                    "description": "本文（任意）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ノート作成成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "タイトルが未入力",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Title is required"
                }
              }
            }
          }
        }
      }
    },
    "/notes/{id}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "summary": "ノート詳細取得",
        "description": "指定IDのノートを1件取得します。認証不要です。",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ノートID"
          }
        ],
        "responses": {
          "200": {
            "description": "ノート詳細",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "404": {
            "description": "ノートが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Note not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Notes"
        ],
        "summary": "ノート更新",
        "description": "指定IDのノートを更新します。Bearer認証が必要です。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ノートID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "タイトル（任意）"
                  },
                  "content": {
                    "type": "string",
                    "description": "本文（任意）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ノート更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "404": {
            "description": "ノートが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Note not found"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notes"
        ],
        "summary": "ノート削除",
        "description": "指定IDのノートを削除します（ソフトデリート）。Bearer認証が必要です。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ノートID"
          }
        ],
        "responses": {
          "200": {
            "description": "ノート削除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Note deleted"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ノートが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Note not found"
                }
              }
            }
          }
        }
      }
    },
    "/payment/checkout": {
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "Stripe Checkout Session を作成",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "priceId"
                ],
                "properties": {
                  "priceId": {
                    "type": "string",
                    "description": "Stripe Price ID"
                  },
                  "quantity": {
                    "type": "integer",
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout Session URL を返却",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "sessionId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "パラメータ不正"
          }
        }
      }
    },
    "/payment/subscription": {
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "サブスクリプション用 Checkout Session を作成",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "priceId"
                ],
                "properties": {
                  "priceId": {
                    "type": "string",
                    "description": "Stripe Price ID（recurring）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout Session URL を返却"
          }
        }
      }
    },
    "/payment/portal": {
      "post": {
        "tags": [
          "Payment"
        ],
        "summary": "Stripe Customer Portal セッションを作成",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Customer Portal URL を返却"
          }
        }
      }
    },
    "/payment/history": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "決済履歴を取得",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "決済履歴一覧"
          }
        }
      }
    },
    "/payment/status/{sessionId}": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "決済ステータスを確認",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "決済ステータス"
          }
        }
      }
    },
...
  },
...
}